+++ changelog.md
@@ -6,6 +6,7 @@
 
 - Updated dev dependencies.
 - Better error message when `execute` middleware option `override` returns not an object.
+- Ensure only desired `execute` middleware options apply to the GraphQL `execute` function.
 - Moved JSDoc type defs into `src/index.js`.
 - Consistent JSDoc syntax style for array types.
 - Moved an `execute` middleware constant from function to module scope.